Change graph representation to be compact
More specifically, when removing a node, we swap_remove it; and the node
labels always go from 0 to num_nodes-1. We need this for the modular
decomposition algorithm, which will be called after removing a node.
This commit implements the basic new logic and most of the so far
implemented algorithms are adjusted, however, more testing is required.