More dynamic graph methods + track inverted labels
Add a bunch of methods to add and remove nodes and edges to the graphs.
We also keep track of the inverted labels in the AdjGraph which only
introduces a fairly small overhead to remove_node but makes
add_labelled_node much more efficient (however, note that this might be
changed in the future, or rather we add an alternative implementation of
AdjGraph, because at the moment I am not sure whether we really need
add_labelled_node for the simulations; we are only going to need a
method to create subgraphs by adding notes (for small subgraphs; for big
ones, it is better to remove nodes), and while this can now be easily
done with add_labelled_node, it could have also been  implemented like
the from_adjancencies methods before this commit).