Make ImplGraph more compatible
The whole idea of the ImplGraph trait was to have a shared graph API so
that the algorithms can be written generically. While this works now -
at least for the two graph data structures we're currently using - it's
a mess (will be cleaned up a little bit as good as possible) and not
really performant (because different graph data structures just behave
really differently). This is not completely unexpected though, and
although I'm going to code up the final simulations with specialised
algorithms for a fixed graph data structure, we are still going to use
the generic versions for testing, because they are easier to check for
correctness.