Refactor the graph module
Prepare the module to be separated into the generic version and the
specialised version. The generic version will focus on clarity and
simplicity in the algorithms, as well as information-rich return values.
It will be used to test the specialised version. The specialised version
will fully focus on performance for our specific use case.