Clean up the `int` mess
Indices (`Node`) are now fixed to be usizes. I'm not sure whether this
is worse or better performance wise; if they would be smaller, we
occasionally would have a better cache line efficiency, however, at the
cost of an additional register mov (e.g., `mov eax, edx`) when indexing.
It is definitely more ergonomic though.
Labels (`Label`) are fixed to be `int`, which is a fixed integer - u32
at the moment - and we require that `int` fits into usize, so that we can
easily cast them via `as` (which does not happen often, but we still we
to be able to do this efficiently).